requirements.txt
asgiref==3.11.0
bcc==0.1.10
diff-match-patch==20241021
dj-database-url==3.1.0
Django==6.0.1
django-cors-headers==4.9.0
django-import-export==4.4.0
djangorestframework==3.16.1
gunicorn==23.0.0
import-export==0.3.1
numpy==2.4.1
packaging==25.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.5
tablib==3.9.0
traitlets==5.14.3
traittypes==0.2.3
whitenoise==6.11.0
